refactor: SketchHandWriterのロジックをカスタムフックに切り出す
